requirements.txt
appdirs==1.4.4
asn1crypto==1.4.0
astroid==2.5.3
attrs==20.3.0
awswrangler==2.6.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
black==20.8b1
boto3==1.17.47
botocore==1.20.47
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
cryptography==3.4.7
cycler==0.10.0
distlib==0.3.1
distro==1.5.0
docker==5.0.0
docker-compose==1.29.0
dockerpty==0.4.1
docopt==0.6.2
et-xmlfile==1.0.1
filelock==3.0.12
flake8==3.9.0
google-pasta==0.2.0
idna==2.10
importlib-metadata==3.10.0
iniconfig==1.1.1
isort==5.8.0
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
llvmlite==0.36.0
lxml==4.6.3
matplotlib==3.4.1
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
numba==0.53.1
numpy==1.20.2
openpyxl==3.0.7
packaging==20.9
pandas==1.2.3
paramiko==2.7.2
pathspec==0.8.1
pep517==0.10.0
pg8000==1.18.0
Pillow==8.2.0
pip-tools==6.0.1
pluggy==0.13.1
protobuf==3.15.7
protobuf3-to-dict==0.1.5
py==1.10.0
pyarrow==3.0.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
pyinstrument==3.4.1
pyinstrument-cext==0.2.4
pylint==2.7.4
PyMySQL==1.0.2
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.3
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
PyYAML==5.4.1
redshift-connector==2.0.877
regex==2021.4.4
requests==2.25.1
s3transfer==0.3.6
sagemaker==2.33.0
scikit-learn==0.24.1
scipy==1.6.2
scramp==1.2.2
shap==0.39.0
six==1.15.0
sklearn-pandas==2.1.0
slicer==0.0.7
smdebug==1.0.5
smdebug-rulesconfig==1.0.1
soupsieve==2.2.1
stepfunctions==2.0.0
texttable==1.6.3
threadpoolctl==2.1.0
toml==0.10.2
tox==3.23.0
tox-gh-actions==2.5.0
tqdm==4.60.0
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.4.3
websocket-client==0.58.0
wrapt==1.12.1
xgboost==1.3.3
zipp==3.4.1
